+++ package/tsup.config.ts
@@ -1,20 +1,22 @@
 import { defineConfig } from "tsup";
-import { peerDependencies } from "./package.json";
+import { peerDependencies, dependencies, devDependencies } from "./package.json";
 
 export default defineConfig((options) => {
 	const dev = !!options.watch;
 	return {
-		entry: ["src/**/*.(ts|js)"],
+		entry: ["src/index.ts"],
 		format: ["esm"],
 		target: "node18",
-		bundle: false,
+		bundle: true,
 		dts: true,
 		sourcemap: true,
 		clean: true,
 		splitting: false,
 		minify: !dev,
+		noExternal: Object.keys(devDependencies),
 		external: [
 			...Object.keys(peerDependencies),
+			...Object.keys(dependencies),
 			/node_modules/g,
 		],
 		tsconfig: "tsconfig.json",