+++ docs/CHANGELOG.md
@@ -1,5 +1,28 @@
 # docs
 
+## 0.0.1-beta.2
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [641dfce]
+- Updated dependencies [7f93ad2]
+  - @domain-expansion/astro@0.1.0-beta.3
+
+## 0.0.1-beta.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @domain-expansion/astro@0.1.0-beta.2
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [095445d]
+  - @domain-expansion/astro@0.1.0-beta.1
+
 ## 0.0.2-beta.0
 
 ### Patch Changes