+++ package/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @domain-expansion/astro
 
+## 0.1.0-beta.3
+
+### Minor Changes
+
+- Change bundle config
+- 641dfce: Make cache configurable
+- 7f93ad2: Change in-memory cache to an LRU
+
+## 0.1.0-beta.2
+
+### Minor Changes
+
+- Sync versions
+
+## 0.1.0-beta.1
+
+### Minor Changes
+
+- 095445d: Cache invalidation
+
 ## 0.1.0-beta.0
 
 ### Minor Changes